integrate(7134): verify live v2 actor authority path
Exercise the persistent three-role MCP workflow through v2 artifact persistence and verification, while splitting workflow tests and projection fixtures into bounded modules.\n\nConstraint: The independent Rust artifact verifier remains the follow-up scope in #7135.\nRejected: Extend this issue into the Rust verifier migration | #7134 explicitly limits that downstream authority consumer.\nConfidence: high\nScope-risk: narrow\nDirective: Keep the live integration test on service envelopes and final projection receipts, not handcrafted actor artifacts.\nTested: node --experimental-strip-types --test .pi/extensions/kamn-mvp/*.test.ts (59 passed); node --check fake MCP fixture modules\nNot-tested: external live service probe\n\nCo-authored-by: OmX <omx@oh-my-codex.dev>
